cheri: riscv: Replace invalid uses of uintptr_t
Some of the riscv code uses uintptr_t for an address sized integer.
With CHERI this is no longer correct. Use unsigned long or __ptraddr_t
instead.

Skip instances in arch/riscv/mm/init.c which needs a much more thorough
rewrite.

+++ arch/riscv/include/asm/cacheinfo.h
@@ -14,7 +14,7 @@ struct riscv_cacheinfo_ops {
 };
 
 void riscv_set_cacheinfo_ops(struct riscv_cacheinfo_ops *ops);
-uintptr_t get_cache_size(u32 level, enum cache_type type);
-uintptr_t get_cache_geometry(u32 level, enum cache_type type);
+unsigned long get_cache_size(u32 level, enum cache_type type);
+unsigned long get_cache_geometry(u32 level, enum cache_type type);
 
 #endif /* _ASM_RISCV_CACHEINFO_H */
